Cleanup xunit
